% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kannisto_thatcher_model.R
\name{calculate_kt_params}
\alias{calculate_kt_params}
\title{Calculate Kannisto-Thatcher extension parameters}
\usage{
calculate_kt_params(lt, age_start_range, fn)
}
\arguments{
\item{lt}{\code{\link[=data.table]{data.table()}} Life table with columns \code{age_start}, \code{mx}, and \code{qx}.}

\item{age_start_range}{\code{\link{numeric(2)}} vector of lower and upper inclusive
age group start bounds to calculate parameters from.}

\item{fn}{Optimization function passed to \code{\link[=optim]{optim()}}.}
}
\value{
\code{numeric(2)} a named vector of optimal parameters \code{a} and \code{b}.
}
\description{
Calculate optimal parameters \eqn{\alpha} and \eqn{\beta} in the
Kannisto-Thatcher model for use in extending a single life table.
}
\details{
\subsection{Parameter selection}{

Initial parameters are derived from the model:

\deqn{\text{logit}(m_x) \sim \ln(\alpha) + \beta * \text{age}}

Optimization is then done using \code{\link[=optim]{optim()}} to minimize the function \code{fn}.
\code{fn} must have the method signature \code{fn(params, data)}, where:
\itemize{
\item \code{params} is a named vector of the parameters \eqn{\alpha} (\code{a}) and
\eqn{\beta} (\code{b})
\item \code{data} is a life table similar to \code{lt}
}
}
}
