% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_qx.R
\name{check_solution}
\alias{check_solution}
\title{Check qx Optimization Solution}
\usage{
check_solution(dt, solution_qx_col, by_vars)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Collection of life tables from a single location, with columns representing
the betas and optomized qx solution.}

\item{solution_qx_col}{[\code{character(1)}]\cr
Column name of \code{dt} containing the optimized qx solution. There is expected
to be only one unique value by group.}

\item{by_vars}{[\code{character()}]\cr
Column names in \code{dt} uniquely identifying a life table.}
}
\value{
[\code{data.table()}]\cr
Table including \code{by_vars} and columns for each condition checked.
}
\description{
Create a table checking different optimization solution conditions for each
grouping in the data.
}
\details{
The conditions checked are:
\enumerate{
\item Is the solution qx >= the qx of the preceding age group?
\item Is the solution qx < 1?
\item Is the implied terminal qx < 1?
}
}
