% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_open_age.R
\name{collapse_open_age}
\alias{collapse_open_age}
\title{Collapse open age of life table}
\usage{
collapse_open_age(lt, open_age, catch_lower_input = FALSE)
}
\arguments{
\item{lt}{\code{data.table()} representing a single life table to (possibly)
collapse.}

\item{open_age}{New open age interval start.}

\item{catch_lower_input}{\code{logical()} determining if an error should be thrown
when the input life table's open age is lower than the given open age.}
}
\value{
Life table with new open age interval, or original life table if the
original open age is lower than the given open age.
}
\description{
If the given open age is less that the current open age of the life table,
collapse deaths and population counts for higher age groups into the new
open age.
}
