% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kannisto_thatcher_model.R
\name{extend_qx_kt}
\alias{extend_qx_kt}
\title{Extend qx with the Kannisto-Thatcher model}
\usage{
extend_qx_kt(lt, id_cols, age_start_range, open_age_start = 110)
}
\arguments{
\item{lt}{\code{\link[=data.table]{data.table()}} Life table with columns \code{age_start}, \code{mx}, \code{qx}, and
\code{id_cols} identifying a single life table.}

\item{id_cols}{\code{\link[=character]{character()}} vector of column names identifying a single
life table}

\item{age_start_range}{\code{\link{numeric(2)}} vector of lower and upper inclusive
age group start bounds to calculate model parameters from.}

\item{open_age_start}{\code{\link{numeric(1)}} start of the open age interval to
predict out to.}
}
\value{
\code{\link[=data.table]{data.table()}} life table(s) with \eqn{q_x} extended out to
\code{open_age_start}.
}
\description{
Extend qx with the Kannisto-Thatcher model
}
