% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_age.R
\name{get_open_age}
\alias{get_open_age}
\title{Get open age interval of a life table}
\usage{
get_open_age(lt)
}
\arguments{
\item{lt}{\code{data.table()} representing a single life table.}
}
\value{
Start of open age interval
}
\description{
Get open age interval of a life table
}
