% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTableProcessing-package.R
\docType{package}
\name{lifeTableProcessing-package}
\alias{lifeTableProcessing}
\alias{lifeTableProcessing-package}
\title{lifeTableProcessing: Functions for Creating Life Tables from Data}
\description{
Life tables can be constructed from mortality and population data. This package implements functions to process that data into life tables under different methods and assumptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ihmeuw-demographics.github.io/lifeTableProcessing/}
}

}
\author{
\strong{Maintainer}: Spencer Pease \email{spease@uw.edu}

}
\keyword{internal}
