% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmd.R
\name{load_raw_hmd_files}
\alias{load_raw_hmd_files}
\title{Load raw HMD files}
\usage{
load_raw_hmd_files(dir, ...)
}
\arguments{
\item{dir}{Path do directory of raw location-specific HMD files}

\item{...}{Additional parameters passed to \code{\link[readr:read_table]{readr::read_table()}}}
}
\value{
\code{\link[=data.table]{data.table()}} compiled from all location files in \code{dir}.
}
\description{
Load raw HMD files
}
