% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kannisto_thatcher_model.R
\name{model_kt_qx}
\alias{model_kt_qx}
\title{Kannisto-Thatcher model in probability of death space}
\usage{
model_kt_qx(x, params)
}
\arguments{
\item{x}{age group start}

\item{params}{\code{\link{numeric(2)}} vector of Kannisto-Thatcher model parameters
\eqn{\alpha} (a) and \eqn{\beta} (b).}
}
\value{
Prediction of Kannisto-Thatcher model at age x
}
\description{
\deqn{
q_x = 1 - (\frac{\alpha * \exp(\beta x) + 1}{\alpha * \exp(\beta * (x+5)) + 1})^{1/\beta}
}
}
