% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lifetable.R
\name{plot_lifetable}
\alias{plot_lifetable}
\title{Plot a life table}
\usage{
plot_lifetable(lt, open_age_var = NULL)
}
\arguments{
\item{lt}{Life table to plot. Should have at least columns \code{loc_id}, \code{year},
\code{sex}, \code{age_start}, and \code{qx}.}

\item{open_age_var}{Optional column name indicating column with the input
open age to use as the color aesthetic.}
}
\value{
a plot of the life table.
}
\description{
Create a log probability of death plot of the given life table, optionally
including multiple lines to indicate different input open age intervals.
}
\examples{
plot_lifetable(usa_2000_processed_lt, "input_open_age")
}
