% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kannisto_thatcher_model.R
\name{predict_kt_qx}
\alias{predict_kt_qx}
\title{Predict probability of death with the Kannisto-Thatcher model}
\usage{
predict_kt_qx(params, base_age, open_age_start = 110)
}
\arguments{
\item{params}{\code{\link{numeric(2)}} vector of Kannisto-Thatcher model parameters
\eqn{\alpha} (a) and \eqn{\beta} (b).}

\item{base_age}{start of age group to begin making predictions from/}

\item{open_age_start}{start of pen age interval (\eqn{q_x = 1}).}
}
\value{
\code{\link[=data.table]{data.table()}} life table of \eqn{q_x} values from \code{base_age} to
the age group preceding the open age interval.
}
\description{
Predict probability of death with the Kannisto-Thatcher model
}
