% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmd.R
\name{prep_hmd}
\alias{prep_hmd}
\alias{prep_hmd_deaths_5x1}
\alias{prep_hmd_pop_5x1}
\alias{prep_hmd_period_lt_5x1}
\title{Prepare HMD data}
\usage{
prep_hmd_deaths_5x1(dir_input)

prep_hmd_pop_5x1(dir_input)

prep_hmd_period_lt_5x1(dir_input_male, dir_input_female)
}
\arguments{
\item{dir_input}{Path to directory containing raw location-specific files for
all sexes.}

\item{dir_input_male}{Path to directory containing raw location-specific
files for males.}

\item{dir_input_female}{Path to directory containing raw location-specific
files for females.}
}
\value{
\code{\link[=data.table]{data.table()}} of HMD data for all locations.
}
\description{
Compile "by statistic" raw data from the \link[=https://mortality.org/]{Human Mortality database}
}
\details{
Currently there are functions to prep data in 5-year age groups and single
year time periods for:
\itemize{
\item deaths: \code{\link[=prep_hmd_deaths_5x1]{prep_hmd_deaths_5x1()}}
\item population: \code{\link[=prep_hmd_pop_5x1]{prep_hmd_pop_5x1()}}
\item life tables: \code{\link[=prep_hmd_period_lt_5x1]{prep_hmd_period_lt_5x1()}}
}
}
