% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmd.R
\name{prep_hmd_age}
\alias{prep_hmd_age}
\title{Prepare age information from HMD data}
\usage{
prep_hmd_age(dt, age_col)
}
\arguments{
\item{dt}{\code{\link[=data.table]{data.table()}} of HMD data.}

\item{age_col}{name of column representing age}
}
\value{
Invisibly returns \code{dt} with \code{age_col} removed in place of \code{age_start} and
\code{age_end}.
}
\description{
Prepare age information from HMD data
}
