% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_qx.R
\name{solve_qx}
\alias{solve_qx}
\title{Solve for optimal input terminal qx}
\usage{
solve_qx(dt, ax_params)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
A single life table to optimize.}

\item{ax_params}{[\code{data.table()}]\cr
Table of ax regression parameters for caclulating ax.}
}
\value{
[\code{data.table()}]\cr
A copy of \code{dt} with a new column appended with the qx solution.
}
\description{
Optimize input terminal qx for one life table based on given terminal mx and
age pattern.
}
