% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kannisto_thatcher_model.R
\name{sse_kt_qx}
\alias{sse_kt_qx}
\title{Calculate sum of squared-errors for qx Kannisto-Thatcher model}
\usage{
sse_kt_qx(params, data)
}
\arguments{
\item{params}{\code{\link{numeric(2)}} vector of Kannisto-Thatcher model parameters
\eqn{\alpha} (a) and \eqn{\beta} (b).}

\item{data}{\code{\link[=data.table]{data.table()}} Life table with at columns \code{age_start} and \code{qx}.}
}
\value{
Sum of square errors of \eqn{q_x}.
}
\description{
\deqn{
\text{SSE} = \sum ((q_x - q_{x,\text{pred}})^2)
}
}
